README.md
wufnix-base: centos7.3-4.min || f27.min;

the idea is to implement all the upto date mining software and drivers specific to the networked interface i will be using into its own spin of one of the ladder distros, then progressively stream lining the interfaces of the distros to a smaller and smaller version of only the required componenets. in the end trying to end up with something similiar to the idea in arch linux. the minimal packages required to build from. while sticking to a "up-from-base standard" base will extend all the other system configurations, like

wufnix-base: miner; gpu miner 
                     firewall; obviously
                     router;  system implemented only to route traffic
                     builder; system used to build <items>, items: <name, relativity>
                     dev; for writting software, fundementelly used by a workstation to do most fo its own stuff without acctualy being on the system
                     workstation; a work station with a graphical user enviroment while keeping the base userinterface and possibly light
                     database; a db for handling  a <item>
                     web-interface; and some more obvious stuff

Honestly i was bored and decided maybe it was worth acctually using my git to put my ideas